Get it to run remotely
